move host/relay to host/autorelay
