+++ p2p/host/autorelay/addrsplosion.go
@@ -1,4 +1,4 @@
-package relay
+package autorelay
 
 import (
 	"encoding/binary"
