+++ p2p/host/autorelay/addrsplosion_test.go
@@ -1,4 +1,4 @@
-package relay
+package autorelay
 
 import (
 	"testing"
