+++ p2p/host/autorelay/autorelay.go
@@ -1,4 +1,4 @@
-package relay
+package autorelay
 
 import (
 	"context"
