+++ p2p/host/autorelay/autorelay_test.go
@@ -1,4 +1,4 @@
-package relay_test
+package autorelay_test
 
 import (
 	"context"
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/libp2p/go-libp2p"
-	"github.com/libp2p/go-libp2p/p2p/host/relay"
+	"github.com/libp2p/go-libp2p/p2p/host/autorelay"
 	relayv1 "github.com/libp2p/go-libp2p/p2p/protocol/circuitv1/relay"
 
 	"github.com/libp2p/go-libp2p-core/event"
@@ -27,8 +27,8 @@ import (
 
 // test specific parameters
 func init() {
-	relay.BootDelay = 1 * time.Second
-	relay.AdvertiseBootDelay = 100 * time.Millisecond
+	autorelay.BootDelay = 1 * time.Second
+	autorelay.AdvertiseBootDelay = 100 * time.Millisecond
 }
 
 // mock routing
@@ -133,16 +133,18 @@ func TestAutoRelay(t *testing.T) {
 	// this is the relay host
 	// announce dns addrs because filter out private addresses from relays,
 	// and we consider dns addresses "public".
-	relayHost, err := libp2p.New(libp2p.DisableRelay(), libp2p.AddrsFactory(func(addrs []ma.Multiaddr) []ma.Multiaddr {
-		for i, addr := range addrs {
-			saddr := addr.String()
-			if strings.HasPrefix(saddr, "/ip4/127.0.0.1/") {
-				addrNoIP := strings.TrimPrefix(saddr, "/ip4/127.0.0.1")
-				addrs[i] = ma.StringCast("/dns4/localhost" + addrNoIP)
+	relayHost, err := libp2p.New(
+		libp2p.DisableRelay(),
+		libp2p.AddrsFactory(func(addrs []ma.Multiaddr) []ma.Multiaddr {
+			for i, addr := range addrs {
+				saddr := addr.String()
+				if strings.HasPrefix(saddr, "/ip4/127.0.0.1/") {
+					addrNoIP := strings.TrimPrefix(saddr, "/ip4/127.0.0.1")
+					addrs[i] = ma.StringCast("/dns4/localhost" + addrNoIP)
+				}
 			}
-		}
-		return addrs
-	}))
+			return addrs
+		}))
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -160,7 +162,7 @@ func TestAutoRelay(t *testing.T) {
 		t.Fatal(err)
 	}
 	relayDiscovery := discovery.NewRoutingDiscovery(relayRouting)
-	relay.Advertise(ctx, relayDiscovery)
+	autorelay.Advertise(ctx, relayDiscovery)
 
 	// the client hosts
 	h1, err := libp2p.New(libp2p.EnableRelay())
