+++ p2p/host/autorelay/relay.go
@@ -1,4 +1,4 @@
-package relay
+package autorelay
 
 import (
 	"context"
